third_party/WebKit/LayoutTests/editing/spelling/focusing-other-frame.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/util.js"></script>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<pre id="console"></pre>
<iframe id="frame1" src="data:text/html,<body contenteditable></body>"></iframe>
<iframe id="frame2" src="data:text/html,<body contenteditable></body>"></iframe>

<script>

description("Last word written in an editable in one frame should be " +
"spellchecked when focusing other frame. To test manually type some " +
"misspelled word in one frame and focus other frame. The word should be marked.");

if (window.internals) {
    internals.settings.setAsynchronousSpellCheckingEnabled(false);

    var frame1 = document.getElementById("frame1");
    var testEditable1 = frame1.contentWindow.document.createElement("div");
    testEditable1.setAttribute("contentEditable", "true");
    frame1.contentWindow.document.body.appendChild(testEditable1);
    var frame2 = document.getElementById("frame2");
    var testEditable2 = frame1.contentWindow.document.createElement("div");
    testEditable2.setAttribute("contentEditable", "true");
    frame2.contentWindow.document.body.appendChild(testEditable2);
    testEditable1.focus();
    frame1.contentWindow.document.execCommand("InsertText", false, "zz");
    shouldBe('internals.markerCountForNode(testEditable1.childNodes[0], "spelling")', '0');
    testEditable2.focus();
    shouldBe('internals.markerCountForNode(testEditable1.childNodes[0], "spelling")', '1');
} else {
    log("Automatic testing impossible. Test manually. See steps in the description.");
}
</script>
</body>
</html>
